+++ Common/src/test/java/testmod/tests/ReplacementsTests.java
@@ -25,6 +25,16 @@ public class ReplacementsTests {
         }
         """, JsonObject.class);
 
+    private static final JsonObject EMPTY_PLACEHOLDERS = JsonUtils.readFromString("""
+        {
+            "type": [],
+            "material": [
+                "iron",
+                "gold"
+            ]
+        }
+        """, JsonObject.class);
+
     @SimpleGameTest
     public void testInflate() {
         var placeholderConfig = PlaceholderConfig.SERIALIZER.handleDeserialization(INFLATE_PLACEHOLDERS);
@@ -37,5 +47,15 @@ public class ReplacementsTests {
         assertTrue(result.contains("c:rods/gold"));
         assertTrue(result.contains("c:raw_materials/iron"));
         assertTrue(result.contains("c:raw_materials/gold"));
+    }
+
+    @SimpleGameTest
+    public void testInflateWithEmptyPlaceholder() {
+        var placeholderConfig = PlaceholderConfig.SERIALIZER.handleDeserialization(EMPTY_PLACEHOLDERS);
+        var result = placeholderConfig.apply("c:{material}");
+
+        assertEquals(2, result.size());
+        assertTrue(result.contains("c:iron"));
+        assertTrue(result.contains("c:gold"));
     }
 }